simplified_v2 may not have been failing because it was int instead of float

+++ sunpy/image/tests/test_transform.py
@@ -367,7 +367,7 @@ def test_minimal_example_simplified():
 
 
 def test_minimal_example_simplified_v2():
-    x = np.arange(300000)
+    x = np.arange(300000, dtype=np.float64)
     src = np.vstack((x, -x, np.ones_like(x))).T
 
     matrix = np.array([[0, -1, 511], [1, 0, 0], [0, 0, 1]]).T
